updated module structure
